feat: monorepo phase 2 -- drift-sdk capability package (types + models, ADR-100)
(cherry picked from commit 6f97a26352151136cfff856ae3d9b923372b027d)
