+++ packages/drift-sdk/src/drift_sdk/models/__init__.py
@@ -198,10 +198,10 @@ __all__ = [
     "PatchStatus",
     "BlastRadius",
     "DiffMetrics",
-    "AcceptanceResult",
     "PatchIntent",
     "PatchVerdict",
+    "AcceptanceResult",
     # Policy
-    "PolicyRule",
     "CompiledPolicy",
+    "PolicyRule",
 ]
